refactor(domain,ci): split waste policies per file, pin CI action SHAs
