+++ docs/en/testing.md
@@ -36,7 +36,7 @@ Pure unit tests, no mocks. One spec per entity, asserting the exposed fields, th
 - [`libs/cloud-cost/domain/src/entities/underutilized-ec2-instance.entity.spec.ts`](../../libs/cloud-cost/domain/src/entities/underutilized-ec2-instance.entity.spec.ts)
 - [`libs/cloud-cost/domain/src/entities/rds-underutilized-instance.entity.spec.ts`](../../libs/cloud-cost/domain/src/entities/rds-underutilized-instance.entity.spec.ts)
 
-Policies live in one file, [`libs/cloud-cost/domain/src/policies/resource-waste-policies.spec.ts`](../../libs/cloud-cost/domain/src/policies/resource-waste-policies.spec.ts), covering for every policy: waste vs not-waste, the grace period, the `ignoreTag`, `excludeTagValues`, and the exact boundary of each threshold (age `===` `minAgeDays`, CPU `===` threshold, ops `===` `maxOps`) — boundaries matter because grace-period and CPU/ops comparisons use opposite-strictness operators (`<` vs `>=`/`>`), so an off-by-one there silently changes which resources get flagged.
+Each policy lives in its own file under [`libs/cloud-cost/domain/src/policies/`](../../libs/cloud-cost/domain/src/policies/) (named after the entity it judges, e.g. `ebs-volume.policy.ts`), mirroring the one-file-per-entity layout above. Their tests are consolidated in one spec, [`libs/cloud-cost/domain/src/policies/waste-policies.spec.ts`](../../libs/cloud-cost/domain/src/policies/waste-policies.spec.ts), covering for every policy: waste vs not-waste, the grace period, the `ignoreTag`, `excludeTagValues`, and the exact boundary of each threshold (age `===` `minAgeDays`, CPU `===` threshold, ops `===` `maxOps`) — boundaries matter because grace-period and CPU/ops comparisons use opposite-strictness operators (`<` vs `>=`/`>`), so an off-by-one there silently changes which resources get flagged.
 
 ### Infra — scanners
