+++ docs/it/test.md
@@ -36,7 +36,7 @@ Unit test puri, senza mock. Uno spec per entità, che verifica i campi esposti, 
 - [`libs/cloud-cost/domain/src/entities/underutilized-ec2-instance.entity.spec.ts`](../../libs/cloud-cost/domain/src/entities/underutilized-ec2-instance.entity.spec.ts)
 - [`libs/cloud-cost/domain/src/entities/rds-underutilized-instance.entity.spec.ts`](../../libs/cloud-cost/domain/src/entities/rds-underutilized-instance.entity.spec.ts)
 
-Le policy vivono in un unico file, [`libs/cloud-cost/domain/src/policies/resource-waste-policies.spec.ts`](../../libs/cloud-cost/domain/src/policies/resource-waste-policies.spec.ts), che copre per ogni policy: waste vs non-waste, il periodo di grazia, l'`ignoreTag`, `excludeTagValues`, e il boundary esatto di ogni soglia (età `===` `minAgeDays`, CPU `===` soglia, ops `===` `maxOps`) — i boundary contano perché periodo di grazia e confronti CPU/ops usano operatori con stretta opposta (`<` contro `>=`/`>`), quindi un errore di uno cambia silenziosamente quali risorse vengono segnalate.
+Ogni policy vive nel proprio file sotto [`libs/cloud-cost/domain/src/policies/`](../../libs/cloud-cost/domain/src/policies/) (chiamato come l'entità che giudica, es. `ebs-volume.policy.ts`), sullo stesso pattern un-file-per-entità visto sopra. I loro test sono consolidati in un unico spec, [`libs/cloud-cost/domain/src/policies/waste-policies.spec.ts`](../../libs/cloud-cost/domain/src/policies/waste-policies.spec.ts), che copre per ogni policy: waste vs non-waste, il periodo di grazia, l'`ignoreTag`, `excludeTagValues`, e il boundary esatto di ogni soglia (età `===` `minAgeDays`, CPU `===` soglia, ops `===` `maxOps`) — i boundary contano perché periodo di grazia e confronti CPU/ops usano operatori con stretta opposta (`<` contro `>=`/`>`), quindi un errore di uno cambia silenziosamente quali risorse vengono segnalate.
 
 ### Infra — scanner
