+++ libs/cloud-cost/infrastructure/aws-adapter/src/scanners/scanner-contract.spec.ts
@@ -23,7 +23,7 @@ import {
   Ec2InstanceWastePolicy,
   EbsSnapshotWastePolicy,
   NatGatewayWastePolicy,
-  Gp2UpgradePolicy,
+  EbsGp2UpgradePolicy,
   EbsIdlePolicy,
   Ec2UnderutilizedPolicy,
   RdsUnderutilizedPolicy,
@@ -206,7 +206,7 @@ const scannerFactories: Record<ResourceKind, () => WasteScannerPort> = {
   'ec2-instance': () => new AwsEc2InstanceScanner(pricing, ACCOUNT, new Ec2InstanceWastePolicy(po)),
   'ebs-snapshot': () => new AwsEbsSnapshotScanner(pricing, ACCOUNT, new EbsSnapshotWastePolicy(po)),
   'nat-gateway': () => new AwsNatGatewayScanner(pricing, ACCOUNT, new NatGatewayWastePolicy(po)),
-  'ebs-gp2-upgrade': () => new AwsGp2UpgradeScanner(pricing, ACCOUNT, new Gp2UpgradePolicy(po)),
+  'ebs-gp2-upgrade': () => new AwsGp2UpgradeScanner(pricing, ACCOUNT, new EbsGp2UpgradePolicy(po)),
   'ebs-idle': () => new AwsEbsIdleScanner(pricing, ACCOUNT, new EbsIdlePolicy(po, 0)),
   'log-group': () => new AwsLogGroupScanner(pricing, ACCOUNT, new LogGroupWastePolicy(po)),
   'eni-orphaned': () => new AwsEniOrphanedScanner(ACCOUNT, new OrphanedEniWastePolicy(po)),
